pyproject.toml
[project]
name = "magnetic-tweezers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "pyside6>=6.9.1",
]
